refactor: spread out hook
